Pull SecurityBasedVideoRenderer up a level and fix verification issues.
